+++ test/Audit.IntegrationTest/EntityFrameworkTests.Core.cs
@@ -67,6 +67,89 @@ namespace Audit.IntegrationTest
         }
 
         [Test]
+        public void Test_EFDataProvider_AuditEntityDisabled_Fluent()
+        {
+            Audit.Core.Configuration.Setup()
+                .UseEntityFramework(config => config
+                    .AuditTypeMapper(typeName => Type.GetType(typeName + "Audit"))
+                    .AuditEntityAction((ev, ent, audEnt) =>
+                    {
+                        return false;
+                    })
+                );
+
+            var id = Guid.NewGuid().ToString();
+            using (var ctx = new AuditPerTableContext())
+            {
+                var o = new Order()
+                {
+                    Number = id,
+                    Status = "Pending",
+                    OrderLines = new Collection<Orderline>()
+                    {
+                        new Orderline() { Product = "p1: " + id, Quantity = 2 },
+                        new Orderline() { Product = "p2: " + id, Quantity = 3 }
+                    }
+                };
+                ctx.Add(o);
+                ctx.SaveChanges();
+            }
+
+            using (var ctx = new AuditPerTableContext())
+            {
+                var order = ctx.Order.Single(a => a.Number.Equals(id));
+                var orderlineAudits = ctx.OrderlineAudit.AsNoTracking().Where(a => a.OrderId.Equals(order.Id)).ToList();
+                Assert.AreEqual(0, orderlineAudits.Count);
+            }
+        }
+        [Test]
+        public void Test_EFDataProvider_AuditEntityDisabled()
+        {
+            var dp = new EntityFrameworkDataProvider();
+            dp.AuditTypeMapper = t =>
+            {
+                if (t == typeof(Order))
+                    return typeof(OrderAudit);
+                if (t == typeof(Orderline))
+                    return typeof(OrderlineAudit);
+                return null;
+            };
+
+            dp.AuditEntityAction = (ev, entry, obj) =>
+            {
+                // return false to avoid saving
+                return false;
+            };
+
+            Audit.Core.Configuration.Setup()
+                .UseCustomProvider(dp);
+            var id = Guid.NewGuid().ToString();
+            using (var ctx = new AuditPerTableContext())
+            {
+                var o = new Order()
+                {
+                    Number = id,
+                    Status = "Pending",
+                    OrderLines = new Collection<Orderline>()
+                    {
+                        new Orderline() { Product = "p1: " + id, Quantity = 2 },
+                        new Orderline() { Product = "p2: " + id, Quantity = 3 }
+                    }
+                };
+                ctx.Add(o);
+                ctx.SaveChanges();
+            }
+
+            using (var ctx = new AuditPerTableContext())
+            {
+                var order = ctx.Order.Single(a => a.Number.Equals(id));
+                var orderlineAudits = ctx.OrderlineAudit.AsNoTracking().Where(a => a.OrderId.Equals(order.Id)).ToList();
+                Assert.AreEqual(0, orderlineAudits.Count);
+            }
+
+        }
+
+        [Test]
         public void Test_EFDataProvider()
         {
             var dp = new EntityFrameworkDataProvider();
@@ -88,6 +171,7 @@ namespace Audit.IntegrationTest
                     ab.UserName = ev.Environment.UserName;
                     ab.AuditStatus = entry.Action; 
                 }
+                return true;
             };
 
             Audit.Core.Configuration.Setup()
@@ -182,6 +266,7 @@ namespace Audit.IntegrationTest
                     ab.UserName = ev.Environment.UserName;
                     ab.AuditStatus = entry.Action;
                 }
+                return true;
             };
 
             Audit.Core.Configuration.Setup()